backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg
alembic==1.12.1
pydantic==2.5.0
httpx==0.25.2
python-dateutil==2.8.2
structlog==23.2.0
bcrypt==4.0.1
PyJWT==2.8.0
python-multipart==0.0.6
redis==5.0.1
pandas==2.1.3
psutil==5.9.6
pytest==7.4.3
pytest-asyncio==0.21.1